feature(ui): add suffix  (#45)
Add suffix UI, using the same configuration method as for the main display formatting.

Also add annotation-function, to provide the same UI pre-Emacs 28..

To enable this, I have adapted some code from bibtex-completion. I will
remove that code, or greatly simplify it, if and when this PR is merged:

https://github.com/tmalsburg/helm-bibtex/pull/367

Also, change 'bibtex-actions-icons' to 'bibtex-actions-symbols'.

Closes #44