refactor: use display-format for cands (#51)
This simplifies the code and user config story, adds matching on
search substrings, and gives us wider compatibility.

Specifically, it:

1. removes the annotation function and affixation suffix
2. adds "suffix" candidate string with the 'bibtex-actions-suffix' face

The 'bibtex-actions--get-candidates' function no longer uses
'bibtex-completion-candidates' for the candidate strings, and instead
uses 'bibtex-actions--format-entry', so that if the user explicitly
configures the UI, they also customize what will be searched.

Closes #47