fix: normalize bibliography file paths where needed
We can't rely on the calls to `citar-file--normalize-paths` in the
default values for the citar-bibliography defcustom, since those will
be overridden by the user. Instead, normalize the paths where needed
in the code.

The same applies to the defcustoms for citar-{library,notes}-paths;
these are always used in contexts where no normalization is needed.

Fixes #351