Autoload keymaps required for embark and simplify readme (#355)
Set up the Embark integration automatically when Citar is loaded, and
also add it to the autoload file. This means that Embark and Citar can
be loaded in any order, and neither will pull in the other. Citar will
be loaded the first time its target finder is invoked by embark-act.

This PR also includes the simplified use-package stanza for citar-org
made possible by #353, allowing its loading to be deferred to the last
possible moment (fixes #340).