Define keymaps as defvar
Using defvars instead of defcustoms is the commonly used style. Keymaps are
customized by modifying the keymaps in place and not by replacing the entire
keymap.

Adding an autoload to a defvar/defcustom is problematic since it copies the
entire definition to the autoload file. Since the target finder is autoloaded
it is not necessary to mark the maps as autoloaded.