Install embark general maps as parent keymap
This change is questionable. I know that you dislike that the general Embark
actions are always there since they clutter which-key. Unfortunately without
this change Embark is badly broken. We need at least the C-h=embark-keymap-help
action. Furthermore in the minibuffer we would still like to have the
keybindings for embark-export/collect-snapshot etc.