Fix embark autoloads
With commit ec0695ba, the citar keymaps are no longer autoloaded and
therefore cannot be referred to by the autoloaded with-eval-after-load
form for embark. This causes an error at emacs initialization.

This change splits the embark with-eval-after-load into two parts: the
first is autoloaded and adds the target finder, while the second deals
with the keymaps. This should be fine because embark will never need
the citar keymaps unless it has (1) run the citar target finder,
or (2) been invoked from an embark command.