docs: simplify use-package block for org
The current configuration causes issues when `use-package-always-ensure` or
`straight-use-package-by-default` are enabled, because they assume that
`citar-org` is a separate package.  Instead, just go with `use-package citar`.

Also remove the `:after org` directive.  Instead use the `:package` argument to
`:bind` to defer the modification of `org-mode-map` until after `org` is
loaded.

Fix #380.