citar-latex-insert-citation: implement default cite style behaviour
new variable: citar-latex-default-cite-command, sets default citation
command for latex, and used as a default in prompting for a citation
command.

new variable: citar-latex-prompt-for-cite-style, sets whether to
prompt for a citation style when inserting.

citar-latex-citation-command: new option invert-prompt. If t, invert
whether or not to prompt for a citation style (as initially determined
by `citar-latex-prompt-for-cite-style'). In actual use, the value of
this argument is inherited from the prefix arg of
`citar-insert-citation'.