citar-file.el: Refactor file-finding functions (#460)
Split the functionality of `citar-file--possible-names` into two new functions:

- 'citar-file--parse-file-field' to extract file names from an entry field
- 'citar-file--directory-files' to find files in directories whose names start
with keys.

Change `citar-file--files-for-entry`, `citar-file--files-for-multiple-entries`,
and `citar-file--get-note-filename` to use these functions.

Add the function `citar-file--has-file-p`, which scans the relevant directories
once and returns a predicate for the existence of associated files. Use this
function in `citar--format-candidates` to efficiently add "has:files" and
"has:notes" indicators.