Improve file parsing.
* `citar-file--parser-default` will now only split file fields by semicolons, not
colons. The colon-separated fields seen in #283 are actually triplet-format
fields, and are handled by `citar-file--parser-triplet`.

* `citar-file--parse-file-field` will now print messages if it fails for any of
the following reasons:
 - The file field exists but is empty.
 - The file field cannot be parsed; it is non-empty but no parser produces
 output.
 - None of the parsed file names actually exist on disk.
 - None of the parsed files have an extension listed in
 `citar-library-file-extensions`.

* `citar-get-file` no longer filters for `citar-library-file-extensions`. This
is now the responsibility of the functions in `citar-get-files-functions`, if
appropriate.

* `citar--library-file-action`, which does the work for `citar-open-files` and
`citar-attach-files`, will only print the "no associated files" warning if none
of its input keys had files according to `citar-has-files`. Otherwise, if a key
was supposed to have files but turned out not to, the appropriate function in
`citar-get-files-functions` should produce a message explaining why.