API: resource getter functions now return hash tables.
The `citar-get-files`, `citar-get-links`, and `citar-get-notes` functions now
return hash tables mapping each key to a list of associated resources. They
used to just return lists of resources for all the given keys, but that
interface doesn't allow callers to determine which key corresponds to each
resource.

The `:items` function in the notes source API has also been modified, and must
now return a hash table instead of a list of notes.