`citar-open` and `citar-open-notes` now allow creating notes.
Both these functions now offer to create notes for the chosen keys. The
customization variable `citar-open-always-create-note` controls whether the
"create" option is offered only for keys that don't already have notes.

This new feature is implemented as an extra option to `citar--select-resource`,
which has now learned about the `create-note` pseudo-resource type. The
`citar-open-notes` and `citar--open-resource` also learned to create new notes.

Other Changes

There is a new command called `citar-open-note`, which prompts to select a note
from all the available notes and opens it. It can also be called
non-interactively with a note returned by `citar-get-notes`, which is its
primary purpose.

The customization variable `citar-open-prompt` can now be set to a list of
commands that will always prompt before selecting resources. The default is to
always prompt for `citar-open`, `citar-attach-files`, and `citar-open-note`.