fix: properly detect the case when there was no input
One way of exiting multiple selection session is to
exit minibuffer with empty input. However when there
are default values the empty input get replaced with
the first default.

Since we were tracking input, one way to detect no
input is to use the tracked input. However this caused
a glaring bug because now user can't select a candidate
without inputting something first.

This commit instead uses the `require-match` argument
of completing-read to detect the case of empty input.
The only case `require-match` function isn't called is
when the empty input is replaced by the default.