fix(org): point is left on the wrong reference after shifting
* citar-org.el: in `citar-org--shift-reference', update citation-reference list
before retrieving the reference's index and moving point

The prior commit changed `citar-org--get-ref-index' to use the begin of
references for equality tests. However, this results in the point being unmoved
after calling `citar-org--shift-reference'. This is because we try to find the
reference's index in the list of references prior to shifting. Thus, we must
call `citar-org--get-ref-index' with an updated list of references to preserve
the point.