fix(org): edge case for citar-org-shift-reference-*
* citar-org.el: test equality with the begin property rather than citekey in
`citar-org--get-ref-index'

Using citekeys to test equality gives false positives because there may be
multiple references with the same citekey. Thus, we should elect to use the
begin property of cite-references, since those are unique.

* citar-org.el: in `citar-org--shift-reference', update citation-reference list
before retrieving the reference's index and moving point

The prior commit changed `citar-org--get-ref-index' to use the begin of
references for equality tests. However, this results in the point being unmoved
after calling `citar-org--shift-reference'. This is because we try to find the
reference's index in the list of references prior to shifting. Thus, we must
call `citar-org--get-ref-index' with an updated list of references to preserve
the point.

* citar-org.el: in `citar-org--shift-reference', record the offset of the point
then apply it after shifting

Prior behavior had point end on the beginning of the reference, regardless of
where the shift was called. This commit allows the point to be left at the same
position relative to the reference it was called on.