feat: Make citar-{bibliography,{library,notes}-paths} safe local vars
Putting the `safe-local-variable` attribute on a variable makes it so that it
can be customized in `.dir-locals.el`, in the `-*- ... -*-` header line, or in a
`Local Variables` block at the end of a file without the user getting a prompt
whether or not to allow this.

`citar-bibliography`, `citar-library-paths`, and `citar-notes-paths` all are
ripe for per-directory customization. For instance, when working on a paper with
collaborators, you might store a bibtex file specific to your project adjacent
to the paper. Suppose your bibliography file is called "custom_refs.bib"; you
can put

    % Local Variables:
    % citar-bibliography: ("./custom_refs.bib")
    % End:

At the end of the file, and now whenever citar is invoked, it will pull up
references from just that bibtex database and not whatever `citar-bibliography`
is set to globally.