refactor: sexp value-type for citar-display-transform-functions
The value type for the custom variable `citar-display-transform-functions' is
currently set to `list', which is not a proper elisp type. This will raise an
error when attempting to byte-compile the file. Instead, the correct value-type
should be sexp.

Moreover, as there nothing is being evaluated, the backtick quotation should be
replaced with a regular quotation to avoid unexpected expansion of the sexp.